% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVGamma.R
\name{rmvgamma}
\alias{rmvgamma}
\title{Multivariate Gamma Distribution}
\usage{
rmvgamma(mu, alpha, beta, samp_size)
}
\arguments{
\item{mu}{vector of location parameter.}

\item{alpha}{vector of shape parameter.}

\item{beta}{scale parameter.}

\item{samp_size}{number of observations.}

\item{x}{matrix of observations from a multivariate gamma distribution, columnar.}

\item{log}{logical; if \code{TRUE}, probabilities are returned as log(p).}
}
\description{
Distribution function and random generator for the multivariate gamma distribution
with parameters \code{mu} for location, \code{alpha} for shape, and \code{beta} for scale. \code{alpha},
\code{x}, and \code{beta} should be in \code{R+}. \code{mu} resides in \code{R}.
}
\examples{
samp<-rmvgamma(mu = c(-1,2), alpha = c(.1,2), beta = 3, samp_size = 10)
samp
dmvgamma(x = samp, mu = c(-1,2), alpha = c(1,1), beta = 3, log = FALSE)
}
