% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StickBreaking.R
\name{rsb}
\alias{rsb}
\title{Random Stick Breaking Process Generator}
\usage{
rsb(alpha, n)
}
\arguments{
\item{alpha}{Concentration parameter in \code{R+}.}

\item{n}{Sample size (integer).}

\item{prior_probs}{a vector of prior probabilities.}
}
\value{
\code{rsb} returns a list, a vector of sampled probabilities
and a vector of probabilities at each draw

\code{rsb_post} returns a list, a vector of sampled probabilities
and a vector of probability at each draw
}
\description{
Random generator for stick breaking process, and posterior stick breaking process.
}
\examples{
samp<-rsb(n=5,alpha=5)
plot(1:5,samp$probs,type = "h")
probs <- c(.2,.2,.6)
samp<-rsb_post(alpha = 5, prior_probs = probs)
plot(1:3, probs,type = "h", col = alpha("red",.4), lwd = 5, xlim =c(0.5,3.5))
lines(1:3+.1, samp$probs,type = "h", col = alpha("blue", .4), lwd = 5)
samp$probs

samp_size<-20
probs<-rsb(alpha = .2,n = samp_size)$probs
samp<-rsb_post(alpha = .5,prior_probs = probs)
plot(1:samp_size, probs,type = "h", col = alpha("red",.4), lwd = 5, xlim =c(0.5,samp_size+.5))
lines(1:samp_size+.1, samp$probs,type = "h", col = alpha("blue", .4), lwd = 5
}
